Fix SqlServer bulkinsert issues breaking the integration tests.
This issue was caused by a combination of a recent changes to keepIdentity handling and some legacy handling that automatically
enables keepIdentity whenever a table has an identity value.

As there was no way to override the behavior it was no longer possible to autogenerate identity values, as we (and probably
our users) do in many scenarios. This caused > 20 test failures.

As there is no easy way to unbreak this I removed the automatically setting keep-identity and cleaned up the arguments to
properly allow setting the value as normal Enum (without nullable wrapper). At the same time I introduced ILogger like
in the non-bulk api and forwarded that wherever possible (and where I did not forget something).

This fixes all the SqlServer Bulkinsert integrationtests.